# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BigCommerce Bulk Importer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings:
    get:
      operationId: getStoreSettings
      tags: [settings]
      summary: Get persisted store configuration
      responses:
        "200":
          description: Store settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSettings"
    put:
      operationId: saveStoreSettings
      tags: [settings]
      summary: Save and validate store credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreSettingsInput"
      responses:
        "200":
          description: Saved settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSettings"
        "422":
          description: Invalid credentials
  /settings/webhooks:
    get:
      operationId: listWebhooks
      tags: [settings]
      summary: List registered BigCommerce webhooks
      responses:
        "200":
          description: Webhook list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookListResponse"
    post:
      operationId: createWebhook
      tags: [settings]
      summary: Register a new BigCommerce webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookInput"
      responses:
        "201":
          description: Created webhook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
  /settings/webhooks/{webhookId}:
    delete:
      operationId: deleteWebhook
      tags: [settings]
      summary: Delete a BigCommerce webhook
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    StoreSettings:
      type: object
      required: [configured]
      properties:
        configured:
          type: boolean
        storeHash:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        accessTokenMasked:
          type: ["string", "null"]
          description: "Last 4 chars of the access token, rest obscured (e.g. xxxxxxxxxxxx1234)"
        clientSecretConfigured:
          type: ["boolean", "null"]
          description: "Whether a client secret has been saved"
        storeName:
          type: ["string", "null"]
        storeUrl:
          type: ["string", "null"]
        updatedAt:
          type: ["string", "null"]
    StoreSettingsInput:
      type: object
      required: [storeHash, accessToken]
      properties:
        storeHash:
          type: string
        accessToken:
          type: string
        clientId:
          type: string
        clientSecret:
          type: string
    Webhook:
      type: object
      required: [id, scope, destination, is_active, created_at, updated_at]
      properties:
        id:
          type: integer
        scope:
          type: string
        destination:
          type: string
        is_active:
          type: boolean
        created_at:
          type: integer
        updated_at:
          type: integer
    WebhookInput:
      type: object
      required: [scope, destination]
      properties:
        scope:
          type: string
        destination:
          type: string
    WebhookListResponse:
      type: object
      required: [webhooks]
      properties:
        webhooks:
          type: array
          items:
            $ref: "#/components/schemas/Webhook"