# lib/api-zod/src/generated/types/webhookInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BigCommerce Bulk Importer API
 * OpenAPI spec version: 0.1.0
 */

export interface WebhookInput {
  scope: string;
  destination: string;
}